pyproject.toml
[tool.poetry]
name = "goose-discord-webhook"
version = "0.1.0"
description = ""
authors = ["Zachary Juang <zachary822@me.com>"]
readme = "README.md"
packages = [{include = "goose_discord"}]

[tool.poetry.dependencies]
python = "^3.10"
awslambdaric = "^2.0.4"
boto3 = "^1.24.85"
pydantic = "^1.10.2"
requests = "^2.28.1"
PyNaCl = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
show_error_codes = true
plugins = ["pydantic.mypy"]

[tool.isort]
profile = "black"
case_sensitive = false
line_length = 120
filter_files = true

[tool.black]
line_length = 120
